# worker/openapi.yaml
openapi: 3.0.3
info:
  title: E-Ink Board Worker API
  version: "1.0"
  description: |
    API for the Cloudflare Worker backing the EE02 e-ink display boards.

    Two very different kinds of clients call this API:

    - **Admin clients** (you, from a browser, curl, Postman, etc.) — everything
      under `/admin/*` plus `GET /current`. Authenticate with
      `Authorization: Bearer <api_key>`. Get a key by completing passkey
      registration (`/auth/register/*`) or by copying it from the admin UI at
      `/admin` ("Rotate API Key").
    - **Devices** (the ESP32 firmware) — `/device_config`, `/hash`,
      `/image_packed`, `/firmware_bin`. These use a different, HMAC-based
      per-request signature (`X-Device-MAC` / `X-Device-Nonce` /
      `X-Device-Signature`), not the Bearer key. Documented here for
      completeness but you won't normally call these by hand.

    Import this file into Postman/Insomnia/Swagger UI, set the `bearerAuth`
    security value to your API key, and every `/admin/*` request (including
    `GET /current`) is one click away.
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost:8787
        description: Your deployed worker's hostname (or a local `wrangler dev` host:port).
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Your account's API key, minted by /auth/register/verify,
        /auth/login/verify, or POST /admin/keys/rotate.
...
  responses:
    RateLimited:
      description: Rate limit exceeded (see lib/rate-limit.ts — generous fixed windows per IP/MAC/user; failing open on limiter errors)
      content:
        text/plain:
          schema: { type: string }
      headers:
        Retry-After:
          description: Seconds until the current window resets
          schema: { type: integer }
paths:
...
  /auth/register/options:
    post:
      summary: Start passkey registration (creates a new account)
      security: []
      responses:
        '200': { description: OK }
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/register/verify:
    post:
      summary: Complete passkey registration
      description: >-
        The three sharing-key fields are optional and present as a set — only
        when this ceremony's authenticator returned a WebAuthn PRF result (see
        root CLAUDE.md's encrypted-buckets plan). `wrapped_sharing_key` is the
        client's P-256 sharing private key, AES-256-GCM-encrypted under a key
        derived from that PRF output — the Worker never sees the PRF output or
        the raw private key, only this ciphertext.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                attemptId: { type: string }
                response: { type: object }
                sharing_public_key: { type: string, description: base64, raw P-256 public key }
                wrapped_sharing_key: { type: string, description: base64 }
                wrap_nonce: { type: string, description: base64, 12 bytes }
              required: [attemptId, response]
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key: { type: string }
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/login/options:
    post:
      summary: Start passkey login
      security: []
      responses:
        '200': { description: OK }
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/login/verify:
    post:
      summary: Complete passkey login
      description: >-
        Mints a fresh API key on every successful login. The response echoes
        back whatever sharing-key wrap is currently on file for this
        credential (possibly all null, if none has ever been established) so
        the client can unwrap its sharing private key using this exact
        ceremony's PRF output. If the client discovers it needs to backfill a
        wrap (no PRF-wrapped key existed yet, but this ceremony returned PRF
        output), it does so via a separate call to
        PATCH /admin/me/sharing-key using the api_key from this response —
        this ceremony's challenge is single-use and already consumed by the
        time that decision can be made.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                attemptId: { type: string }
                response: { type: object }
              required: [attemptId, response]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key: { type: string }
                  sharing_public_key: { type: string, nullable: true }
                  wrapped_sharing_key: { type: string, nullable: true }
                  wrap_nonce: { type: string, nullable: true }
...